1314: Refactor metrics code to make it easier to maintain
