+++ accessibility_monitoring_platform/apps/common/views.py
@@ -25,15 +25,11 @@ from ..audits.models import (
 )
 from ..cases.models import (
     Case,
-    WEBSITE_INITIAL_COMPLIANCE_COMPLIANT,
     RECOMMENDATION_NO_ACTION,
     ACCESSIBILITY_STATEMENT_DECISION_COMPLIANT,
     CLOSED_CASE_STATUSES,
 )
-from ..s3_read_write.models import S3Report
-from ..reports.models import ReportVisitsMetrics
 
-from .chart import LineChart, build_yearly_metric_chart
 from .forms import (
     AMPContactAdminForm,
     AMPIssueReportForm,
@@ -42,16 +38,13 @@ from .forms import (
     FrequentlyUsedLinkOneExtraFormset,
 )
 from .metrics import (
-    ThirtyDayMetric,
-    Timeseries,
-    TimeseriesDatapoint,
-    TimeseriesHtmlTable,
-    group_timeseries_data_by_month,
     calculate_metric_progress,
     count_statement_issues,
-    build_html_table,
-    convert_timeseries_pair_to_ratio,
-    convert_timeseries_to_cumulative,
+    get_case_progress_metrics,
+    get_report_progress_metrics,
+    get_case_yearly_metrics,
+    get_policy_yearly_metrics,
+    get_report_yearly_metrics,
 )
 from .models import FrequentlyUsedLink, IssueReport, Platform, ChangeToPlatform
 from .page_title_utils import get_page_title
@@ -60,7 +53,6 @@ from .utils import (
     get_platform_settings,
     record_model_update_event,
     record_model_create_event,
-    get_days_ago_timestamp,
 )
 
 
@@ -210,87 +202,10 @@ class MetricsCaseTemplateView(TemplateView):
     def get_context_data(self, **kwargs) -> Dict[str, Any]:
         """Add number of cases to context"""
         context: Dict[str, Any] = super().get_context_data(**kwargs)
-        thirty_days_ago: datetime = get_days_ago_timestamp()
-        sixty_days_ago: datetime = get_days_ago_timestamp(days=60)
-
-        progress_metrics: List[ThirtyDayMetric] = [
-            ThirtyDayMetric(
-                label="Cases created",
-                last_30_day_count=Case.objects.filter(
-                    created__gte=thirty_days_ago
-                ).count(),
-                previous_30_day_count=Case.objects.filter(created__gte=sixty_days_ago)
-                .filter(created__lt=thirty_days_ago)
-                .count(),
-            ),
-            ThirtyDayMetric(
-                label="Tests completed",
-                last_30_day_count=Case.objects.filter(
-                    testing_details_complete_date__gte=thirty_days_ago
-                ).count(),
-                previous_30_day_count=Case.objects.filter(
-                    testing_details_complete_date__gte=sixty_days_ago
-                )
-                .filter(testing_details_complete_date__lt=thirty_days_ago)
-                .count(),
-            ),
-            ThirtyDayMetric(
-                label="Reports sent",
-                last_30_day_count=Case.objects.filter(
-                    report_sent_date__gte=thirty_days_ago
-                ).count(),
-                previous_30_day_count=Case.objects.filter(
-                    report_sent_date__gte=sixty_days_ago
-                )
-                .filter(report_sent_date__lt=thirty_days_ago)
-                .count(),
-            ),
-            ThirtyDayMetric(
-                label="Cases closed",
-                last_30_day_count=Case.objects.filter(
-                    completed_date__gte=thirty_days_ago
-                ).count(),
-                previous_30_day_count=Case.objects.filter(
-                    completed_date__gte=sixty_days_ago
-                )
-                .filter(completed_date__lt=thirty_days_ago)
-                .count(),
-            ),
-        ]
-
-        yearly_metrics: List[
-            Dict[
-                str,
-                Union[str, TimeseriesHtmlTable, List[TimeseriesDatapoint], LineChart],
-            ]
-        ] = []
-        now: datetime = django_timezone.now()
-        start_date: datetime = datetime(now.year - 1, now.month, 1, tzinfo=timezone.utc)
-        for label, date_column_name in [
-            ("Cases created", "created"),
-            ("Tests completed", "testing_details_complete_date"),
-            ("Reports sent", "report_sent_date"),
-            ("Cases completed", "completed_date"),
-        ]:
-            timeseries: Timeseries = Timeseries(
-                label=label,
-                datapoints=group_timeseries_data_by_month(
-                    queryset=Case.objects,
-                    date_column_name=date_column_name,
-                    start_date=start_date,
-                ),
-            )
-            yearly_metrics.append(
-                {
-                    "label": f"{label} over the last year",
-                    "html_table": build_html_table(columns=[timeseries]),
-                    "chart": build_yearly_metric_chart(lines=[timeseries]),
-                }
-            )
 
         extra_context: Dict[str, Any] = {
-            "progress_metrics": progress_metrics,
-            "yearly_metrics": yearly_metrics,
+            "progress_metrics": get_case_progress_metrics(),
+            "yearly_metrics": get_case_yearly_metrics(),
         }
         return {**extra_context, **context}
 
@@ -403,129 +318,12 @@ class MetricsPolicyTemplateView(TemplateView):
             ),
         ]
 
-        thirteen_month_retested_audits: QuerySet[Audit] = Audit.objects.filter(
-            retest_date__gte=thirteen_month_start_date
-        )
-        thirteen_month_website_initial_compliant: QuerySet[
-            Audit
-        ] = thirteen_month_retested_audits.filter(
-            case__website_compliance_state_initial=WEBSITE_INITIAL_COMPLIANCE_COMPLIANT
-        )
-        thirteen_month_statement_initial_compliant: QuerySet[
-            Audit
-        ] = thirteen_month_retested_audits.filter(
-            case__accessibility_statement_state=ACCESSIBILITY_STATEMENT_DECISION_COMPLIANT
-        )
-        thirteen_month_final_no_action: QuerySet[
-            Audit
-        ] = thirteen_month_retested_audits.filter(
-            case__recommendation_for_enforcement=RECOMMENDATION_NO_ACTION
-        )
-        thirteen_month_statement_final_compliant: QuerySet[
-            Audit
-        ] = thirteen_month_retested_audits.filter(
-            case__accessibility_statement_state_final=ACCESSIBILITY_STATEMENT_DECISION_COMPLIANT
-        )
-
-        retested_by_month: Timeseries = Timeseries(
-            label="Cases",
-            datapoints=group_timeseries_data_by_month(
-                queryset=thirteen_month_retested_audits,
-                date_column_name="retest_date",
-                start_date=thirteen_month_start_date,
-            ),
-        )
-        website_initial_compliant_by_month: Timeseries = Timeseries(
-            label="Initially acceptable",
-            datapoints=group_timeseries_data_by_month(
-                queryset=thirteen_month_website_initial_compliant,
-                date_column_name="retest_date",
-                start_date=thirteen_month_start_date,
-            ),
-        )
-        statement_initial_compliant_by_month: Timeseries = Timeseries(
-            label="Initially compliant",
-            datapoints=group_timeseries_data_by_month(
-                queryset=thirteen_month_statement_initial_compliant,
-                date_column_name="retest_date",
-                start_date=thirteen_month_start_date,
-            ),
-        )
-        final_no_action_by_month: Timeseries = Timeseries(
-            label="Finally acceptable",
-            datapoints=group_timeseries_data_by_month(
-                queryset=thirteen_month_final_no_action,
-                date_column_name="retest_date",
-                start_date=thirteen_month_start_date,
-            ),
-        )
-        statement_final_compliant_by_month: Timeseries = Timeseries(
-            label="Finally compliant",
-            datapoints=group_timeseries_data_by_month(
-                queryset=thirteen_month_statement_final_compliant,
-                date_column_name="retest_date",
-                start_date=thirteen_month_start_date,
-            ),
-        )
-
-        website_initial_ratio: Timeseries = convert_timeseries_pair_to_ratio(
-            label="Initial",
-            partial_timeseries=website_initial_compliant_by_month,
-            total_timeseries=retested_by_month,
-        )
-        website_final_ratio: Timeseries = convert_timeseries_pair_to_ratio(
-            label="Final",
-            partial_timeseries=final_no_action_by_month,
-            total_timeseries=retested_by_month,
-        )
-        statement_initial_ratio: Timeseries = convert_timeseries_pair_to_ratio(
-            label="Initial",
-            partial_timeseries=statement_initial_compliant_by_month,
-            total_timeseries=retested_by_month,
-        )
-        statement_final_ratio: Timeseries = convert_timeseries_pair_to_ratio(
-            label="Final",
-            partial_timeseries=statement_final_compliant_by_month,
-            total_timeseries=retested_by_month,
-        )
-
-        yearly_metrics: List[Dict[str, Union[str, TimeseriesHtmlTable, LineChart]]] = [
-            {
-                "label": "Proportion of websites which are acceptable",
-                "html_table": build_html_table(
-                    columns=[
-                        retested_by_month,
-                        website_initial_compliant_by_month,
-                        final_no_action_by_month,
-                    ],
-                ),
-                "chart": build_yearly_metric_chart(
-                    lines=[website_initial_ratio, website_final_ratio],
-                    y_axis_percent=True,
-                ),
-            },
-            {
-                "label": "Proportion of accessibility statements which are compliant",
-                "html_table": build_html_table(
-                    columns=[
-                        retested_by_month,
-                        statement_initial_compliant_by_month,
-                        statement_final_compliant_by_month,
-                    ],
-                ),
-                "chart": build_yearly_metric_chart(
-                    lines=[statement_initial_ratio, statement_final_ratio],
-                    y_axis_percent=True,
-                ),
-            },
-        ]
-
         extra_context: Dict[str, Any] = {
             "total_metrics": total_metrics,
             "progress_metrics": progress_metrics,
             "equality_body_cases_completed_count": equality_body_cases_completed_count,
             "equality_body_cases_in_progress_count": equality_body_cases_in_progress_count,
-            "yearly_metrics": yearly_metrics,
+            "yearly_metrics": get_policy_yearly_metrics(),
         }
         return {**extra_context, **context}
 
@@ -540,83 +338,10 @@ class MetricsReportTemplateView(TemplateView):
     def get_context_data(self, **kwargs) -> Dict[str, Any]:
         """Add number of cases to context"""
         context: Dict[str, Any] = super().get_context_data(**kwargs)
-        thirty_days_ago: datetime = get_days_ago_timestamp()
-        sixty_days_ago: datetime = get_days_ago_timestamp(days=60)
-
-        progress_metrics: List[ThirtyDayMetric] = [
-            ThirtyDayMetric(
-                label="Published reports",
-                last_30_day_count=S3Report.objects.filter(created__gte=thirty_days_ago)
-                .filter(latest_published=True)
-                .count(),
-                previous_30_day_count=S3Report.objects.filter(
-                    created__gte=sixty_days_ago,
-                )
-                .filter(created__lt=thirty_days_ago)
-                .filter(latest_published=True)
-                .count(),
-            ),
-            ThirtyDayMetric(
-                label="Report views",
-                last_30_day_count=ReportVisitsMetrics.objects.filter(
-                    created__gte=thirty_days_ago
-                ).count(),
-                previous_30_day_count=ReportVisitsMetrics.objects.filter(
-                    created__gte=sixty_days_ago,
-                )
-                .filter(created__lt=thirty_days_ago)
-                .count(),
-            ),
-            ThirtyDayMetric(
-                label="Reports acknowledged",
-                last_30_day_count=Case.objects.filter(
-                    report_acknowledged_date__gte=thirty_days_ago
-                ).count(),
-                previous_30_day_count=Case.objects.filter(
-                    report_acknowledged_date__gte=sixty_days_ago,
-                )
-                .filter(report_acknowledged_date__lt=thirty_days_ago)
-                .count(),
-            ),
-        ]
-
-        now: datetime = django_timezone.now()
-        start_date: datetime = datetime(now.year - 1, now.month, 1, tzinfo=timezone.utc)
-        published_reports_by_month: Timeseries = Timeseries(
-            label="Published reports",
-            datapoints=group_timeseries_data_by_month(
-                queryset=S3Report.objects.filter(latest_published=True),
-                date_column_name="created",
-                start_date=start_date,
-            ),
-        )
-        report_views_by_month: Timeseries = Timeseries(
-            label="Report views",
-            datapoints=group_timeseries_data_by_month(
-                queryset=ReportVisitsMetrics.objects,
-                date_column_name="created",
-                start_date=start_date,
-            ),
-        )
-
-        yearly_metrics: List[Dict[str, Union[str, TimeseriesHtmlTable, LineChart]]] = [
-            {
-                "label": "Reports published over the last year",
-                "html_table": build_html_table(columns=[published_reports_by_month]),
-                "chart": build_yearly_metric_chart(
-                    lines=[convert_timeseries_to_cumulative(published_reports_by_month)]
-                ),
-            },
-            {
-                "label": "Reports views over the last year",
-                "html_table": build_html_table(columns=[report_views_by_month]),
-                "chart": build_yearly_metric_chart(lines=[report_views_by_month]),
-            },
-        ]
 
         extra_context: Dict[str, Any] = {
-            "progress_metrics": progress_metrics,
-            "yearly_metrics": yearly_metrics,
+            "progress_metrics": get_report_progress_metrics(),
+            "yearly_metrics": get_report_yearly_metrics(),
         }
 
         return {**extra_context, **context}
